will add minikube setup in Vagrant VM
1. Why is this change necessary ?

The current Vagrantfile in maya repository helps
in setting up a development environment for maya
code. It helps in compile & building Docker images of
mayactl, apiserver & others.

However, there is a need to have some features tested
before submitting a PR. Some of these features require
a K8s setup to test the functionality. Hence, it
will be ideal to update the current Vagrantfile to
include minikube setup.

2. How does this change address the issue ?

- Modified Vagrantfile to include minikube & kubectl

3. How to verify this change ?

- Run below commands:
  - Vagrant up
  - kubectl get po

4. What side effects does this change have ?

- None