1. Why is this change necessary ?
It has been seen that policy enforcement is required
for OpenEBS volumes during its provisioning operations.
A volume policy can be related to placements, security/backup,
workload characteristics, QOS, etc.

Policy Enforcement Workflow: A volume provisioning request
to maya is intercepted by volume_endpoint.go file. This is
made to pass through various policy enforcement logic. The
end point decides on the volume provisioner based on VolumeType.
It is then directed to the selected provisioner. The provisioner
then decides on the volume orchestrator provider & directs the
request to the selected orchestrator. The orchestrator in turn
is responsible to convert the volume provisioning request to its
own structure and pass it to its APIs.

This particular commit does a partial
fix of openebs/openebs#409, openebs/openebs#721,
openebs/openebs#749, openebs/openebs#751

2. How does this change address the issue ?

- Add & Delete volume goes through policy enforcement logic

- More properties & setters are introduced in K8s structures to
  handle state & in turn make them unit-testable

- Functions have been introduced to create new instances of
  certain structures. This will remove the need for fetching
  instances via registration

- Policy related operations contract has been introduced in
  orchestrator interface

- A new, simple but redundant wrapper over K8s client-go
  operations has been introduced. This avoids modification of
  existing code while caters to new requirements. This will
  be used instead of older wrapper in near future.

- Typed structures have been introduced instead of
  map[string]string

3. How to verify this change ?

- make should work fine

- a set of issues will be created that verifies
  policy enforcement via unit tests, e2e tests

4. What side effects does this change have ?

- Testing is required to un-earth side effects

- Lack of unit tests is a concern