will enable querying the task result property via alias
1/ Why is this change necessary?

This enables executing jsonpath queries without providing
the jsonpath, but by mentioning only the alias. The logic has
a in-memory map of known aliases w.r.t their jsonpath.

e.g. earlier it was as follows:
```yaml
    queries:
    - alias: objectName
      path: |-
        {.metadata.name}
```

now it can be as follows:
```yaml
    queries:
    - alias: objectName
```

2/ How does this change address this issue?

The change was there already. However, earlier it tried
to make the alias value case insensitive. Now the alias
value is case sensitive. This has been done to avoid
confusions.

3/ How can one verify the change?

A sample volume policy template has been provided to
clarify things further

```yaml
apiVersion: v1
kind: ConfigMap
metadata:
  name: volume-pvc-0.6.0
  annotations:
    openebs.io/policy: VolumePolicy
    policy.openebs.io/version: 0.6.0
  namespace: default
data:
  meta: |
    runNamespace: {{ .Volume.runNamespace }}
    apiVersion: v1
    kind: PersistentVolumeClaim
    objectName: {{ .Volume.pvc }}
    action: get
    queries:
    - alias: objectName
      path: |-
        {.metadata.name}
    - alias: affinity
      path: |-
        {.metadata.annotations.controller\.openebs\.io/affinity}
    - alias: affinityTopology
      path: |-
        {.metadata.annotations.controller\.openebs\.io/affinity-topology}
```

- In above volume policy template the task defines a get call to PVC API

- In addition, there are queries to extract specific parts of the result
due to above API call

- Some of the queries have got path i.e. jsonpath instructions

- However, the path is optional if the alias' value is understood
by the logic.

- In this case, alias=objectName is understood by the logic. The logic
sets the path for this internally as {.metadata.name}.

NOTE: In a volume policy template objectName refers to the object being
retrieved, created, etc. In this case its the name of the PVC that is
being retrieved.

4/ Does this change have any side-effects?

- One needs to ensure if the alias has a corresponding path in the
logic. If in doubt, provide the path as well.

- The logic will error out if the alias is not mapped in the
logic & the volumepolicy template skips providing the path as well.

Signed-off-by: AmitKumarDas <amit.das@openebs.io>