Unit testing and delete event handling 1. Why is this change necessary ? fixes: openebs/openebs#1428, openebs/openebs#1429, openebs/openebs#1407
2. How does this change address the issue ?
The delete event handler is implemented differently, since autogen
Get() call fails as the resource is deleted. Before passing the key onto
the workqueue, the object is stored in a cache and retrieved in the
other end of workqueue in case of delete event which is stored and handled
in QueueLoad. Unit testing is written without including pool-volume binaries
and realtime k8s config. The pool/volume imported should need to be stored
in CStorPool controller and checked if a volumereplica event comes. This is to
differentiate imported pool-vol and user trying to create a volume that is
already present(duplication).

3. How to verify this change ?
Apply the cstor deployment at openebs/openebs/k8s/crd-samples.

Signed-off-by: gkGaneshR <gkganesh126@gmail.com>