Change IMAGE_PREFIX to use travis branch tag 1. Why is this change necessary ?    Instead of getting docker image tag from version file, it would be good to make use of TRAVIS_BRANCH.
2. How does this change address the issue ?
Following pseudo code is handled in travis-build.sh, considering
different upstream branches and local branch:
```
if "$TRAVIS_BRANCH" != ""
 if "$TRAVIS_BRANCH" == "$MASTER"
  IMAGE_PREFIX=ci
 else
  IMAGE_PREFIX=$TRAVIS_BRANCH-ci
else
 IMAGE_PREFIX=local
```
This will be made use in docker build and push image prefix(docker tags)

3. How to verify this change ?
It is verified from docker image tags being built and pushed to docker hub
by travis.

Note: From travis doc,
TRAVIS_BRANCH:
for push builds, or builds not triggered by a pull request, this is the name of the branch.
for builds triggered by a pull request this is the name of the branch targeted by the pull request.
for builds triggered by a tag, this is the same as the name of the tag (TRAVIS_TAG).

Signed-off-by: gkGaneshR <gkganesh126@gmail.com>