Change the key of CAS Template
The name of cas template for create/delete should
be present on spc yaml as annotaions.

Annotation is a key-value pair.

Previous setup was following :
[Key]: [Value]
openebs.io/create-template:[Name of the cas template]

This commit will change the Key for create cas template to following:
[Key]: [Value]
cas.openebs.io/create-pool-template: [Name of the cas template]

After merging the commit, the SPC yaml should use the updated key
only for successful creation of pool.

Signed-off-by: sonasingh46 <sonasingh46@gmail.com>