Runtask as a custom resource (#447)
* convert a runtask as a kubernetes custom resource

This commit has following changes:
- pkg/apis/openebs.io/v1alpha1/cas_template.go has appropriate
annotations to generate client go files
- client code to invoke runtask custom resource is generated
at pkg/client/generated

* update runtask as a custom resource than configmap

This commit enables use of runtask as a custom resource
than via a configmap. Logic handles the cases where
environment does not have the runtask as a custom
resource but has older config maps that represents a
runtask.

RunTask has become a custom resource to enable versioning
and hopefully make use of other features provided by
kubernetes to any Custom Resource Definition.

* add pkg/errors are one of the dependency

This commit enables import of pkg/errors as one of the
dependency. This package helps in adding info as the
error bubbles out. This also helps in getting the root
cause when needed and printing error stacktrace when
desired.


Signed-off-by: AmitKumarDas <amit.das@mayadata.io>