Bug fix: runNamespace not found.
Using config tlp for fetching runNamespace was failing.
This was because the default values were not being set
by the Run() method.

The value of c.templateValues[string(v1alpha1.ConfigTLP)]]
would never be nil as it is being set to an array during
cas volume build.

Changed the innitial value of the same to nil.

Signed-off-by: Prince Rachit Sinha <prince.rachit@mayadata.io>