enhance transforming the kind to resource
This commit has following fix:
- Enhances the logic to transform the value of a resource
kind to its corresponding resource value
- e.g. kind 'pod' becomes resource 'pods'
- e.g. kind 'storageclass' becomes resource 'storageclasses'

NOTE:
Installer code makes use of client-go's dynamic package which
in turn needs the use of GroupVersionResource. In addition,
installer code takes yaml as input and tries to convert that
to an unstructured instance before applying the same into
kubernetes cluster. Since installer does not have a mechanism
to get the resource value, it tries to build it based on
kind's value.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>