refactor(cast): remove unused parameters from SPC
cacheFile option provided in the SPC is ignored, and
a unique cacheFile path is auto generated using the SPC uuid
and provided to cstor-pool. The hardcoded value in SPC is
unusable, since it has to be unique across pools.

overProvisioning is always hardcoded to false. Providing
this option in the SPC will given an impression to user
that it can be enabled/disabled. Until this feature is
supported, removed the need to specify it via SPC.

Signed-off-by: kmova <kiran.mova@openebs.io>