fix(runtask): Fix available pool count check
The cstor-volume-create-listcstorpoolcr-default-0.7.0 task was getting
the size of pool as actual count + 1. This is because when the list is
split there is an extra entry with content "" appended to the array.
e.g. 'a;b;' when split will give 3 values instead of two.

This was causing the pool size to appear as one size extra.
Fixed this by adding a function splitListTrim in template.go which would
first remove prefix and suffix seperator and then internally use the
same logic as that of splitList.

Signed-off-by: Prince Rachit Sinha <prince.rachit@mayadata.io>