fix(cast): put the volume annotations on the pod
Refer: https://github.com/openebs/openebs/issues/1907

FSType details were being placed on the deployment annotations
but the read CAST was looking for them from the pod annotations.
This was causing the FSType to be empty in CASVolume spec
returned to provisioner - resulting in ext4 only being used.

Made the change to volume controller deployment CAST to place
the annotations on the pod.

Signed-off-by: kmova <kiran.mova@openebs.io>