feat(cstor-volume-mgmt): Add phase info to cstor volume cr.
The cstor volume lacks any information about the actual state of volume.
Add information about the state of volume i.e. ReadOnly, Degraded,
Running, Error based on the replica status.

Added a new Operation called 'Sync' which will trigger the CV update.

Run informer in a loop so that it can recover in case of panic.

Signed-off-by: princerachit <prince.rachit@mayadata.io>