refactor(cstor-volume-mgmt): Use status being supplied by istgt command.
The istgtcontrol command is expected to return response in a new json
format. This PR uses the updated json format.

The response is also going to contain the overall status of the volume.
If the cv is not in error,init,invalid states then the status in
response would be updated on the CV object.

Signed-off-by: princerachit <prince.rachit@mayadata.io>