feat(install, cast) add kubernetes version as label to default CASTemplates
This commit adds kubernetes version as a label to CASTemplate
e.g. 'kubeVersion: v1.9.7-gke.11'. In addition, a number of version
related template functions have been exposed that can be used during
go templating.

Below are the introduced template functions:
- kubeVersionCompare
- kubeVersionEq
- kubeVersionGt
- kubeVersionGte
- kubeVersionLt
- kubeVersionLte

Below is a sample CASTemplate that gets installed due to this changes:

apiVersion: openebs.io/v1alpha1
kind: CASTemplate
metadata:
  labels:
    castName: storage-pool-read-default-0.8.0
    kubeVersion: v1.9.7-gke.11
    version: 0.8.0
  name: storage-pool-read-default-0.8.0

Setting version as a label in CASTemplate results in having this label
as a part of cast config that can be used via go templating in
mapped RunTasks, for example:
if {{ kubeVersionEq .CAST.kubeVersion "v1.10.1.alpha"}}

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>