fix(monitoring): add default prometheus scrape annotations
For monitoring the OpenEBS volumes, the user had to setup
the prometheus config with a scarping job to collect metrics
from openebs volume target pods. There are multiple ports
accessible, and the job had to drop the irrelevant ports.

This PR helps in setting up the default prometheus
annotations that specify the scrape path and port.

Signed-off-by: kmova <kiran.mova@openebs.io>