fix(cstor_operator): filter SP correctly based on SPC
This commit will do following:

When a pool is created, it is checked that for that SPC no two pools
are created on the same node.
For this,we figure out existing SP for the given SPC by passing SPC name
to the SP list options.

The list options in list of SP was not passed correctly which caused to
give a list of SP which was created due to other SPC.
So this stopped a SPC to create pool on nodes that was used by other SPC
to create pools.

This commit fixes this.

Signed-off-by: Ashutosh Kumar <ashutosh.kumar@openebs.io>