docs(design, specs): adds openebs operator design specifications
This commit adds the design specification document of openebs
operator. This helps a reviewer to understand the thoughts &
design decisions of openebs operator. One of the aims of this
document is to help the reviewer understand openebs operator
code during its review cycles as well as much later when the
code has been merged.

This can be thought of as openebs operator's high level design
document.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>