docs(design, specs): adds openebs operator design specifications
This specifications is a simple to understand document that can be
refered to during code reviews of openebs operator. One should be
able to easily identity various points mentioned here with multiple
usecases that it tries to implement. However, this is not a use-case
document. It tries to manage the expectations of code reviewers and
personas who want to understand/skim through the technical aspects
minus any significant learning curve. In addition, this document
tries **not** to be a deep dive technical journal & would like its
readers to read/review the corresponding code instead.

In other words, I have tried to document the things I would have
explained to someone reviewing the operator code. One can expect
direct phrases like "It is XYZ", "It is **not** ABC", etc while
reading through this document.

**Special notes for your reviewer**:
This can be thought of as openebs operator's high level design
document.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>