enhance(CAST): Adds support for target affinity in statefulset
Signed-off-by: Ashish Ranjan <ashishranjan738@gmail.com>

This commit adds support for target affinity for statefulset applications. In case of STS applications using OpenEBS volumes, target and replica related resources need to have affinity with application. This means, for a given OpenEBS volume, target and replica related resources should be on the same node where application resides.