fix(stringer, debuggability): remove cyclic calls in stringer package (#1084)
### Background
Most of the API structures in maya override String() & GoString() methods of Stringer & GoStringer interfaces. This is done to print details of the the structure while logging errors.

### Why is this PR needed?
String() & GoString() make use of Yaml() to pretty-print the entire instance.
However, in-case of any error during un-marshaling to yaml, this instance is provided as-is to fmt.Sprintf as a fallback approach. However, this leads to invocation of String() once again since instance is an implementation of String() &/or GoString(). This results into a never ending loop causing a panic.

This commit solves above panic.

Signed-off-by: Shovan Maity shovan.maity@mayadata.io