chore(remove used code): remove controller-runtime code
This commit removes code associated with controller-runtime
and openebs cluster operator.

Maya will make use of client-go to build controllers and
operators. Since openebs cluster is not yet developed, it
made sense to remove it now and come back with a design
that makes use of plain client-go to build its controller.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>