feat(usage): ability to include installer type in GA events
OpenEBS can be installed using different modes like helm,
operator, or partners like rancher, nks etc. This PR
helps track which type of installer are being used.

Also, this PR pushes the cluster identifier in an
additional field, that can be used by the reports
to co-relate volumes with clusters.

Signed-off-by: kmova <kiran.mova@mayadata.io>