fix(apis): change ndm apis (#1302)
Changed BlockDevice and BlockDeviceClaim APIs

Introduced a new state Released in BlockDevice ClaimState. A BD will be in Released state once the associated BDC is deleted, and cleanup jobs are being performed on the BD. Once the cleanup job is successful, BD will be marked as Unclaimed and can be claimed by other BDCs

Removed Mountpoint from BDC. The field is irrevelant because it is not possible to create a claim with a particular mount point, since it involves iterating through all the BDs.

BlockVolumeMode is introduced which is used to specify the type of BD to claim. Depending on the volume mode, either unformatted device or BD with a filesystem can be claimed


Signed-off-by: Akhil Mohan <akhil.mohan@mayadata.io>