refactor(kubeclient): add singleton instance to upgraderesult kubeclient
This commit adds function that exposes creating a singleton instance
for upgraderesult's kubeclient. This helps the caller code while
creating a kubeclient instance since most of the properties of a
kubeclient instance need to be set once in the lifetime of the
process.

In addition, fetching a singleton instance of pkg/kubernetes/client
has been modified to include functional options. This singleton
instance has not been used in upgraderesult in this commit, but will
be used with other necessary changes in future commit(s).

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>