feat(install): add flag to disable default config creation
Refer: https://github.com/openebs/openebs/issues/2665

The default storage pool and storage classes installed
by openebs-apiserer can't be customized by the user.

While further thoughts have to go into what aspects of
the storage configuration needs to be customized, this
PR allows to disable the creation of the default
storage classes/pools.

The new logic for creation of default config is as follows:
(a) if default config is enabled, create:
    - jiva storage pool and jiva default storage class
    - local pv storage classes for hostpath and device

(b) if default config is enabled and cstor sparse is enabled:
    - create cstor sparse pool claim and its storage class

By default:
- config creation is enabled.
- cstor sparse is disabled

Note that the openebs snapshot promoter storage class
is installed and managed by openebs apiserver, as there
are no customization supported by this storage class.

Signed-off-by: kmova <kiran.mova@mayadata.io>