fix(jiva): patch replica deployment with correct hostname
Related issue: https://github.com/openebs/openebs/issues/2695

Jiva Replica Deployment uses `nodeAffinity` to ensure that
replica pods are pinned to the nodes where they are initially created.
This is achieved by the following steps during provisioning:
- Launch the ReplicaDeployment without `nodeAffinity` with
  required number of replicas.
- Read the `spec.nodeName` from each of the scheduled replica
  pods.
- Patch the repplica deployment with `nodeAffinity` by
  setting `kubernetes.io/hostname` to be `In` the list
  of selected node names.

It has been observed that in some deployments hostname != nodename.
This results in jiva replicas not getting scheduled.

This PR converts the nodename into hostname before patching
with nodeAffinity. The CAS template function to support
this conversion was added in the PR: https://github.com/openebs/maya/pull/1389

The same coversion is also done in the cleanup job for setting
the nodeSelector.

Note: We explored the alternate mechanism of patching the
nodeAffinity with `matchFields` set directly to `metadata.name`.
However `matchFields` only supports matching against a single value
and can't be used for Replica Deployment that requires matching with
more than 1 value (node names) depending on the replica count.

Signed-off-by: kmova <kiran.mova@mayadata.io>