feat(cstorpool): changes 'cstorpool' cr to 'cstorpoolinstance'
CStorPoolInstance is the custom resources which represents a cStor pool
on a node. It is at namespace scope and under group 'openebs.io' .
The version of the CR is 'v1alpha1' with this commit.
CStorPoolInstance is represented as 'cspi' in short form which can be used in
cli interactions.

CStorPoolInstance( CSPI ) should not be manipulated/edited by users manually.

Signed-off-by: Ashutosh Kumar <ashutosh.kumar@openebs.io>