chore(cstor-volume-mgmt): add knownreplicas information in spec of cstor volume
This PR adds knownreplicas information in spec section of cstor-volume CR.
With this PR we are maintaining trusty/known replica information in spec
as well as status of cstorvolume.

Why we need to maintain in KnownReplicas information in spec
This is needed operator/user will remove known replica information under
spec section of cStorvolume(during replica scaledown scenario).

Why we need to maintain in KnownReplicas information in status
This is needed to known whether scaledown operation is successfull or
not(When scaledown is triggered cstor-volume-mgmt will update status
with remaining replicas).

Signed-off-by: mittachaitu <sai.chaithanya@mayadata.io>